+++ src/server.rs
@@ -1632,6 +1632,16 @@ impl ServerRunner {
         }
     }
 
+    /// Runs the server.
+    ///
+    /// # Errors
+    ///
+    /// This fails when the server cannot be started.
+    #[deprecated = "Use `Self::run_until_interrupt()` instead."]
+    pub fn run(self) -> Result<()> {
+        self.run_until_interrupt()
+    }
+
     /// Runs the server until interrupted.
     ///
     /// The server is shut down cleanly upon receiving the `SIGINT` signal at which point the method
@@ -1640,7 +1650,7 @@ impl ServerRunner {
     /// # Errors
     ///
     /// This fails when the server cannot be started.
-    pub fn run(mut self) -> Result<()> {
+    pub fn run_until_interrupt(mut self) -> Result<()> {
         let Self {
             server,
             access_control_sentinel,
@@ -1689,7 +1699,7 @@ impl ServerRunner {
     /// # Errors
     ///
     /// This fails when the server cannot be started.
-    pub fn run_until_cancelled(mut self, is_cancelled: &mut impl FnMut() -> bool) -> Result<()> {
+    pub fn run_until_cancelled(mut self, mut is_cancelled: impl FnMut() -> bool) -> Result<()> {
         let Self {
             server,
             access_control_sentinel,