+++ CHANGELOG.md
@@ -25,6 +25,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   deprecated in 0.7.1. Use `status()` and/or `with_status()` instead.
 - Breaking: Remove method `ua::String::as_slice()` deprecated in 0.5.0. Use `as_bytes()` instead.
 - Breaking: Bump Minimum Supported Rust Version (MSRV) to 1.83.
+- Rename `ServerRunner::run()` to `ServerRunner::run_until_interrupt()`. Deprecate former method.
 - Mark `AsyncMonitoredItem::into_stream()` as deprecated. Use the `Stream` implementation of this
   type directly instead.
 - Remove `unsafe` qualifier from `Server::statistics()`, it may now be be called concurrently.