+++ CHANGELOG.md
@@ -7,7 +7,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
-## [0.9.0] - 2025-05-30
+## [0.9.0] - 2025-06-02
 
 ### Added